seperated functions for organizational purposes
