+++ pages/job_search.py
@@ -1,6 +1,7 @@
 import json
 from datetime import datetime
 from typing import Optional
+from pages.application import applyToJob
 
 from common_utils.messages import (
     anyButtonToContinueMessage,
@@ -125,79 +126,6 @@ def printJobOptionScreen(jobIndex, currentUser: Optional[User] = None) -> Option
             input("")
 
     return
-
-
-def applyToJob(jobIndex, currentUser):
-    jobs = loadJobs()
-    applicationNum = len(jobs[jobIndex]["applicants"])
-    flag = False
-    # checking to see if currentUser has already applied
-    for i in range(0, applicationNum):
-        if jobs[jobIndex]["applicants"][i]["username"] == currentUser.username:
-            flag = True
-        else:
-            i += 1
-    # If user is not logged in
-    if not isinstance(currentUser, User):
-        print("You must be logged in to create a Job.")
-        print(anyButtonToContinueMessage())
-        input("")
-        return currentUser
-    # If user created the job posting
-    elif (
-        jobs[jobIndex]["firstname"] == currentUser.firstname
-        and jobs[jobIndex]["lastname"] == currentUser.lastname
-    ):
-        print("You cannot apply for a job you posted.")
-        print(anyButtonToContinueMessage())
-        input("")
-        return currentUser
-    elif flag == True:
-        print("You cannot apply for a job you have already applied to.")
-        print(anyButtonToContinueMessage())
-        input("")
-        return currentUser
-    else:
-        clearScreen()
-        while True:
-            print("*** Job Application ***")
-            gradDate = input("Enter graduation date in the form mm/dd/yyy: ")
-            if checkDate(gradDate) == False:
-                print("Invalid input please try again")
-                print(anyButtonToContinueMessage())
-                input("")
-                break
-            startDate = input("Enter your desired start date in the form mm/dd/yyy: ")
-            if checkDate(startDate) == False:
-                print("Invalid input please try again")
-                print(anyButtonToContinueMessage())
-                input("")
-                break
-            explanation = input("Explain why you would be a good fit for this position: ")
-
-            jobs[jobIndex]["applicants"] = [
-                {
-                    "username": currentUser.username,
-                    "graduation": gradDate,
-                    "start": startDate,
-                    "reason": explanation,
-                }
-            ]
-            saveJobDatabase(JSON_JOBS_FP, jobs)
-            print("Application sucessfully submitted")
-            print(anyButtonToContinueMessage())
-            input("")
-            break
-
-    return currentUser
-
-
-def checkDate(input_string):
-    try:
-        datetime.strptime(input_string, "%m/%d/%Y")
-        return True
-    except ValueError:
-        return False
 
 
 def userSaveJob(jobIndex, currentUser):
